rentro/pmp-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic>=2.0
pydantic-settings>=2.0
alembic
email-validator
bcrypt<4.0.0
passlib[bcrypt]
python-jose[cryptography]
boto3
python-multipart
watchfiles
apscheduler
requests>=2.31.0
sendgrid==6.11.0
jinja2==3.1.3
python-dateutil
